SA-8: decompose ProjectSourcingPage
